Box blur and gaussian blur get correctly multiplied by the scalar. kernel is now float[][]

+++ src/DistrKerneling.java
@@ -2,7 +2,7 @@
 
 public class DistrKerneling {
 
-    public void test(String fileName, String directory, int[][] kernel){
+    public void test(String fileName, String directory, float[][] kernel){
 
         System.out.println("Function is not yet implemented, sorry for the inconvenience :(");
