+++ src/ParKerneling.java
@@ -1,6 +1,6 @@
 public class ParKerneling {
 
-    public void test(String fileName, String directory, int[][] kernel){
+    public void test(String fileName, String directory, float[][] kernel){
 
         System.out.println("Function is not yet implemented, sorry for the inconvenience :(");
